+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## [Unreleased]
 
+## [0.0.4] - 23/01/2025
+
+- Enhanced `Context` class with the ability to update thread, messages, and run details.
+- Create DatabaseManagement class to handle database interact function
+- Enable thread worker count customization
+- Enable worker's task timeout customization
+
 ## [0.0.3] - 14/01/2025
 
 - Generate new thread, message or run id based on metadata